msm: kgsl: Don't allocate memory dynamically for temp command buffers
The temporary command buffer in adreno_iommu_set_pt_ctx is only the size of
a single page, and adreno_iommu_set_pt_ctx is never executed concurrently.
It is therefore easy to replace the dynamic command buffer allocation with
a static one to improve performance by avoiding the latency of dynamic
memory allocation.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mrsrimar22 <mrsrimar22@gmail.com>