input: ft3418_i2c: Queue touch suspend work on workqueue
In case of continuous panel BLANK and UNBLANK, there is a possibility
of a race condition in the current sequence of operations. This change
ensures that suspend work is also queued on the same workqueue as the
resume work to ensure no race around and removes the need for having to
cancel touch resume work during a panel blank.

Change-Id: I1e0c64a7cd3e001c7224b2ea357877a45c1116e4
Signed-off-by: ppadasal <quic_ppadasal@quicinc.com>
[fiqri19102002: Port to ft3418_i2c driver]
Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>
Signed-off-by: muralivijay <muralidharan9845@gmail.com>
Signed-off-by: mrsrimar22 <mrsrimar22@gmail.com>